+++ Stackful/src/sfshared.cpp
@@ -8,7 +8,7 @@
 namespace stackful {
 	SFLiteral_p atomNil, atomFalse, atomTrue, atomMissing;
 	std::regex functionArityRegex(R"~((^.[^\/]*?)\/?(\*|\d+)?$)~");
-	SFFunctionArity_t getFunctionArity(const std::string &name, const int np) {
+	SFFunctionArity_t getFunctionArity(const std::string &name, const size_t np) {
 		std::sregex_iterator iter(name.begin(), name.end(), functionArityRegex);
 		std::sregex_iterator end;
 		if (iter == end) {
